+++ packages/@react-native-windows/cli/src/e2etest/templates/cs-lib.test.ts
@@ -5,5 +5,5 @@
  * @format
  */
 
-import snapshotTemplate from './snapshotTemplates';
+import snapshotTemplate from './snapshotTemplate';
 snapshotTemplate('cs', 'lib');
