+++ packages/@react-native-windows/cli/src/e2etest/templates/snapshotTemplate.ts
